fix(extension): move tenant-room field into Settings (always reachable)
The room field was added under the "Signal server" status row, but that
panel only renders when meshStatus==Ready && dkgState==Idle — i.e. hidden
when Offline (first open) and during DKG, exactly when you need to set the
room. Moved it to the Settings panel (the ⚙ gear), which is always
reachable, as a "Signal server room" card: input + Generate (UUID) + Save
(validated). Reverted the App.svelte block/handlers.

svelte-check: 0 errors; extension builds (wxt) cleanly.
